Update CI tag to full tag name
- Now image generated have the full name, if the tag passed is either master, develop, stable
  or latest.
- Currently only the regex corresponding to full name is checked. (see
  example above)

A concrete example:

```
./generate_recipe.sh feelpp/feelpp-toolboxes:latest
./generate_recipe.sh feelpp/feelpp-toolboxes:latest
./generate_push.sh -g <girder_folder_id> feelpp/feelpp-toolboxes:latest
```

This will create a recipe/image for docker image
`feelpp/feelpp-toolboxes:develop-v0.104.0-alpha.3-ubuntu-16.04`
and push it in the <gider_folder_id> folder on the girder server.
If the tag change for latest, the image is automatically changed to last
version, the old release image is kept!